requirements/base.txt
python-dotenv>=0.19.2
authlib==1.2.0
requests>=2.27.1
uvicorn==0.21.1
fastapi==0.92.0
jwt==1.3.1
fastapi_auth0==0.3.2
starlette==0.25.0
httpx==0.23.3
fastapi-sessions==0.3.2
pyyaml==6.0
python-jose==3.3.0
openapi-schema-pydantic==1.2.4
sqlalchemy==2.0.6
pymysql==1.0.2
email-validator==1.3.1
pytest==7.2.2
stomp.py==8.1.0
python-multipart==0.0.6
